+++ dakara_server/dakara_server/settings/development.py
@@ -1,27 +1,50 @@
+"""
+Django local settings for the Dakara server project.
+
+For more information on this file, see
+https://docs.djangoproject.com/en/1.11/topics/settings/
+
+For the full list of settings and their values, see
+https://docs.djangoproject.com/en/1.11/ref/settings/
+
+You should not modify this file directly.
+To modify config values, set them as environment variables,
+or in a config file in the dakara root directory:
+either in a `.env` file
+or in a `settings.ini` with a single `[settings]` section.
+"""
+
 import os
 
-BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+from decouple import config, Csv
+from dj_database_url import parse as db_url
+
+from .base import *  # noqa F403
+from .base import BASE_DIR
 
 # Quick-start development settings - unsuitable for production
 # See https://docs.djangoproject.com/en/1.11/howto/deployment/checklist/
 
-# SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY = 'hc85nuu)6iu0z6!=5_y=0norbd)g)$syxy*9kw+3np@=0x)#k4'
+SECRET_KEY = config('SECRET_KEY', default='YourSecretKey')
+DEBUG = config('DEBUG', cast=bool, default=True)
+ALLOWED_HOSTS = config('ALLOWED_HOSTS', cast=Csv(), default='*')
 
-# SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = True
-
-# SECURITY WARNING: change it to your server hostname in production!
-ALLOWED_HOSTS = ['*']
+# Django password security policy
+# https://docs.djangoproject.com/en/1.11/topics/auth/passwords/#module-django.contrib.auth.password_validation
+AUTH_PASSWORD_VALIDATORS = config('AUTH_PASSWORD_VALIDATORS',
+                                  cast=Csv(), default="")
 
 # Database
 # https://docs.djangoproject.com/en/1.11/ref/settings/#databases
+# `DATABASE_URL` is specified according to dj-databse-url plugin
+# https://github.com/kennethreitz/dj-database-url#url-schema
 
 DATABASES = {
-    'default': {
-        'ENGINE': 'django.db.backends.sqlite3',
-        'NAME': os.path.join(BASE_DIR, 'db.sqlite3'),
-    }
+    'default': config(
+        'DATABASE_URL',
+        default='sqlite:///' + os.path.join(BASE_DIR, 'db.sqlite3'),
+        cast=db_url
+    )
 }
 
 # Channels
@@ -33,13 +56,12 @@ CHANNEL_LAYERS = {
     },
 }
 
-
 # Internationalization
 # https://docs.djangoproject.com/en/1.11/topics/i18n/
 
-LANGUAGE_CODE = 'en-us'
+LANGUAGE_CODE = config('LANGUAGE_CODE', default='en-us')
 
-TIME_ZONE = 'UTC'
+TIME_ZONE = config('TIME_ZONE', default='UTC')
 
 # Loggin config
 LOGGING = {
@@ -95,7 +117,10 @@ LOGGING = {
         },
         'django': {
             'handlers': ['console'],
-            'level': os.getenv('DJANGO_LOG_LEVEL', 'INFO'),
+            'level': config('DJANGO_LOG_LEVEL', default='INFO'),
         },
     },
 }
+
+# limit of the playlist size
+PLAYLIST_SIZE_LIMIT = config('PLAYLIST_SIZE_LIMIT', cast=int, default=100)